+++ test/manual/cache-sw.html
@@ -2,7 +2,7 @@
 <html ⚡>
 <head>
   <meta charset="utf-8">
-  <title>Ad examples</title>
+  <title>Cache SW Manual Testing</title>
   <link rel="canonical" href="http://nonblocking.io/">
   <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
   <script>
@@ -10,8 +10,6 @@
       'cache-service-worker': 1
     };
   </script>
-  <!-- temporary, until new changes make it into production -->
-  <!-- <script async src="https://cdn.ampproject.org/v0.js"></script> -->
   <script async src="../../dist/v0.js"></script>
   <style amp-boilerplate>body{-webkit-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-moz-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-ms-animation:-amp-start 8s steps(1,end) 0s 1 normal both;animation:-amp-start 8s steps(1,end) 0s 1 normal both}@-webkit-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-moz-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-ms-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-o-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}</style><noscript><style amp-boilerplate>body{-webkit-animation:none;-moz-animation:none;-ms-animation:none;animation:none}</style></noscript>
   <script async src="../../dist/v0/amp-accordion-0.1.js"></script>
